exists -> utils and applyToAllFunction
